Move account query into its own file
