--- backend-rust/.sqlx/query-dde8ae171c95f3839d26649e99087ea53c363f5dbfb1562e540a5be48ab38fb6.json
@@ -1,15 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "UPDATE accounts\n            SET amount = amount + $1\n            WHERE address = $2",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Int8",
-        "Text"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "dde8ae171c95f3839d26649e99087ea53c363f5dbfb1562e540a5be48ab38fb6"
-}
